blog/requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11
django-crispy-forms==1.7.0
django-filter==1.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.4
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
Markdown==2.6.11
markdown2==2.3.5
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2017.3
pywin32==222
qrcode==5.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
virtualenv==15.1.0
